fix(automation): unify flow-mode action execution with simple mode (EVO-1641)
AutomationRules::FlowExecutionService reimplemented its conversation
actions standalone and had silently diverged from the simple-mode
executor (ActionService). Extract the conversation-bound actions into a
single shared module (ConversationActionHandlers), included by both
executors, so the two surfaces can no longer diverge — same EVO-1262
mixin pattern as PipelineActionHandlers / MessageActionHandlers.

Fixes the four documented divergences by construction:
- change-status-node: was missing from action_node?/dispatch (silent
  no-op) -> wired to the canonical change_status.
- send_email_transcript: was a logging stub -> now actually sends via
  ConversationReplyMailer (parity with simple mode).
- assign_agent: now goes through the agent_belongs_to_inbox? guard, so a
  flow can no longer assign an agent outside the conversation's inbox.
- webhook event string: unified to automation_event.{event_name} across
  the simple, flow and contact executors (was automation_flow.* /
  contact_*). The string is outbound-only (WebhookJob POSTs it to the
  customer URL); no internal consumer branches on it.

Side effect of the parity: add/remove-label on a conversation now goes
through update!(label_list:), which dirty-tracks the change and creates
the label activity-message in the timeline (the old label_list.add+save!
path skipped it). Labels themselves were always applied correctly; this
mirrors the EVO-1635 F-2 fix already shipped for contacts.

FlowExecutionService now owns only flow-control (walking nodes/edges) and
node_data normalisation — zero action bodies. ActionService / Macros
behaviour is unchanged (specs pass unaltered). Net -368/+358.
